moved register_binary() to main.py
