+++ suite2p/registration/__init__.py
@@ -1,3 +1,4 @@
 from .metrics import get_pc_metrics
-from .register import register_binary, pick_initial_reference
+from .register import pick_initial_reference
+from .main import register_binary
 from .zalign import compute_zpos
